+++ shake/skeletons/project/bundles/common/__init__.py
@@ -14,8 +14,7 @@ from . import views
 
 urls = [
     EndpointPrefix('bundles.common.views', [
-
-        Rule('/', 'index', name='index'), 
-        
+        Rule('/', 'index',
+            name='index'), 
     ]),
 ]