+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Shake Changelog
 
+## Version 1.5
+
+- Added a `wrapper` parameter to `shake.templates.link_to` for making navigations bootstrap-style.
+
+- User roles in the default project skeleton, and other changes.
+
 
 ## Version 1.4